refactor(streams): move PoolAutoscaler to Streams namespace (#323)
PoolAutoscaler is a streams-pool controller, not web-specific — it was
mis-namespaced under Web::Streamer in #327. Move it to Pgbus::Streams so the
Client (Layer 1) can drive it from the publisher path without depending on the
web layer (Layer 6). Pure rename: lib/pgbus/web/streamer/pool_autoscaler.rb →
lib/pgbus/streams/pool_autoscaler.rb, all refs + log tags updated, no behavior
change. Streamer + streams specs green (533).
